+++ src/crab/finite_domain.hpp
@@ -10,15 +10,15 @@
 #include "asm_syntax.hpp" // for Condition::Op
 #include "crab/interval.hpp"
 #include "crab/split_dbm.hpp"
+#include "crab/var_registry.hpp"
 #include "string_constraints.hpp"
-#include "var_registry.hpp"
 
 namespace prevail {
 
 class FiniteDomain {
     SplitDBM dom;
 
-public:
+  public:
     explicit FiniteDomain(const SplitDBM& dom) : dom{dom} {}
 
     explicit FiniteDomain() = default;
@@ -128,9 +128,7 @@ public:
     void sign_extend(Variable svalue, Variable uvalue, const LinearExpression& right_svalue, int target_width,
                      int source_width);
 
-    bool add_constraint(const LinearConstraint& cst) {
-        return dom.add_constraint(cst);
-    }
+    bool add_constraint(const LinearConstraint& cst) { return dom.add_constraint(cst); }
 
     void set(const Variable x, const Interval& intv) { dom.set(x, intv); }
